Revert "feat: made changes for the improvement of the website, getting it ready for launch "
