+++ app/layout.tsx
@@ -4,9 +4,7 @@ import { Inter, Nunito } from 'next/font/google';
 import "./globals.css";
 import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
-import ClientLayout from '@/app/client-layout';
-import Loader from '@/components/loader';
-
+import ClientLayout from '@/app/client-layout'
 
 const nunito = Nunito({
   subsets: ["latin"],
@@ -26,7 +24,6 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body className={nunito.className}>
-        <Loader />
         <ClientLayout>{children}</ClientLayout>
       </body>
     </html>
