--- app/messages/page.tsx
@@ -1,12 +1,0 @@
-import React from 'react'
-import Message from './message'
-
-const page = () => {
-  return (
-    <div>
-        <Message />
-    </div>
-  )
-}
-
-export default page
